use codegen for ValueEntity sample
