+++ .circleci/config.yml
@@ -137,6 +137,29 @@ jobs:
           command: make -C docs
       - save_sbt_cache
 
+  samples-tests:
+    machine: true
+    description: "Maven Samples Integreation Tests"
+    steps:
+      - checkout-and-merge-to-main
+      - setup_sbt
+      - restore_sbt_cache
+      - run: 
+          name: "Publish local maven"
+          command: | 
+            # tag the head with the hash
+            # dynver requires a version number prefix, ie: v0.0.1-
+            git tag v0.0.1-"$CIRCLE_SHA1"
+            echo 'export SDK_VERSION=0.0.1-${CIRCLE_SHA1}' >> $BASH_ENV
+            sbt publishM2
+      - run: 
+          name: Run integration tests for Value Entities sample
+          command: |
+            cd samples/java-valueentity-shopping-cart 
+            echo "Running mvn with syntetic SDK version: '$SDK_VERSION'"
+            mvn -Dakkaserverless-sdk.version=$SDK_VERSION verify -Pit
+      - save_sbt_cache
+
   publish:
     docker:
       - image: circleci/openjdk:11
@@ -189,6 +212,10 @@ workflows:
           filters: # required since publish jobs have tag filters and dependencies
             tags:
               only: /.*/
+      - samples-tests:
+          filters: # required since publish jobs have tag filters and dependencies
+            tags:
+              only: /.*/ 
       - tck-tests:
           filters: # required since publish jobs have tag filters and dependencies
             tags:
@@ -207,8 +234,9 @@ workflows:
             - checks
             - tests
             - integration-tests
+            - samples-tests    
             - tck-tests
-            - validate-docs
+            - validate-docs   
       - publish-tck:
           filters: # version tags only
             tags:
