+++ samples/java-valueentity-shopping-cart/docker-compose.linux.yml
@@ -1,0 +1,5 @@
+version: "3"
+services:
+  akka-serverless-proxy:
+    extra_hosts:
+      - "host.docker.internal:host-gateway"
