+++ samples/java-valueentity-shopping-cart/docker-compose.yml
@@ -1,0 +1,15 @@
+version: "3"
+services:
+  akka-serverless-proxy:
+    image: gcr.io/akkaserverless-public/akkaserverless-proxy:0.7.0-beta.9
+    command: -Dconfig.resource=dev-mode.conf -Dakkaserverless.proxy.eventing.support=google-pubsub-emulator
+    ports:
+      - "9000:9000"
+    environment:
+      USER_FUNCTION_HOST: ${USER_FUNCTION_HOST:-host.docker.internal}
+      USER_FUNCTION_PORT: ${USER_FUNCTION_PORT:-8080}
+  gcloud-pubsub-emulator:
+    image: gcr.io/google.com/cloudsdktool/cloud-sdk:341.0.0
+    command: gcloud beta emulators pubsub start --project=test --host-port=0.0.0.0:8085
+    ports:
+      - 8085:8085
